# docs/api.yaml
openapi: 3.1.0
info:
  title: Student Management API
  version: 0.1.0
paths:
  /auth/register:
    post:
      tags:
      - Auth
      summary: Register User
      operationId: register_user_auth_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegister'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /auth/login:
    post:
      tags:
      - Auth
      summary: Login
      operationId: login_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /users/me:
    get:
      tags:
      - Users
      summary: Read Users Me
      operationId: read_users_me_users_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
      security:
      - OAuth2PasswordBearer: []
  /users/:
    get:
      tags:
      - Users
      summary: List Users
      operationId: list_users_users__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/UserResponse'
                type: array
                title: Response List Users Users  Get
      security:
      - OAuth2PasswordBearer: []
  /students/:
    post:
      tags:
      - Students
      summary: Create Student
      operationId: create_student_students__post
      security:
      - OAuth2PasswordBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Students
      summary: List Students
      operationId: list_students_students__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Search by name or university ID
          title: Search
        description: Search by name or university ID
      - name: department
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Department
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: gpa_min
        in: query
        required: false
        schema:
          anyOf:
          - type: number
            maximum: 4.0
            minimum: 0
          - type: 'null'
          title: Gpa Min
      - name: gpa_max
        in: query
        required: false
        schema:
          anyOf:
          - type: number
            maximum: 4.0
            minimum: 0
          - type: 'null'
          title: Gpa Max
      - name: skip
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Skip
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 10
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StudentResponse'
                title: Response List Students Students  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /students/{student_id}:
    get:
      tags:
      - Students
      summary: Get Student
      operationId: get_student_students__student_id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: student_id
        in: path
        required: true
        schema:
          type: integer
          title: Student Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - Students
      summary: Update Student
      operationId: update_student_students__student_id__patch
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: student_id
        in: path
        required: true
        schema:
          type: integer
          title: Student Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Students
      summary: Delete Student
      operationId: delete_student_students__student_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: student_id
        in: path
        required: true
        schema:
          type: integer
          title: Student Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /students/stats/summary:
    get:
      tags:
      - Students
      summary: Get Student Stats
      description: Get a summary of student statistics.
      operationId: get_student_stats_students_stats_summary_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - OAuth2PasswordBearer: []
  /monitoring/metrics:
    get:
      tags:
      - Monitoring
      summary: Get Metrics
      operationId: get_metrics_monitoring_metrics_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Get Metrics Monitoring Metrics Get
  /monitoring/dashboard:
    get:
      tags:
      - Monitoring
      summary: Get Dashboard
      operationId: get_dashboard_monitoring_dashboard_get
      responses:
        '200':
          description: Successful Response
          content:
            text/html:
              schema:
                type: string
  /health:
    get:
      tags:
      - Health
      summary: Health Check
      description: Check the health of the system with real measurements.
      operationId: health_check_health_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Health Check Health Get
  /:
    get:
      tags:
      - Root
      summary: Read Root
      operationId: read_root__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
                title: Response Read Root  Get
components:
  schemas:
    Gender:
      type: string
      enum:
      - male
      - female
      title: Gender
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    Role:
      type: string
      enum:
      - admin
      - student
      title: Role
    StudentCreate:
      properties:
        university_id:
          type: string
          maxLength: 20
          minLength: 3
          title: University Id
        name:
          type: string
          maxLength: 150
          minLength: 1
          title: Name
        birth_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Birth Date
        gender:
          anyOf:
          - $ref: '#/components/schemas/Gender'
          - type: 'null'
        phone_number:
          anyOf:
          - type: string
            maxLength: 20
          - type: 'null'
          title: Phone Number
        gpa:
          type: number
          maximum: 4.0
          minimum: 0.0
          title: Gpa
          default: 0.0
        department:
          type: string
          maxLength: 100
          minLength: 1
          title: Department
        enrollment_date:
          type: string
          format: date
          title: Enrollment Date
        status:
          $ref: '#/components/schemas/StudentStatus'
          default: active
        user_id:
          type: integer
          title: User Id
      type: object
      required:
      - university_id
      - name
      - department
      - user_id
      title: StudentCreate
    StudentResponse:
      properties:
        university_id:
          type: string
          maxLength: 20
          minLength: 3
          title: University Id
        name:
          type: string
          maxLength: 150
          minLength: 1
          title: Name
        birth_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Birth Date
        gender:
          anyOf:
          - $ref: '#/components/schemas/Gender'
          - type: 'null'
        phone_number:
          anyOf:
          - type: string
            maxLength: 20
          - type: 'null'
          title: Phone Number
        gpa:
          type: number
          maximum: 4.0
          minimum: 0.0
          title: Gpa
          default: 0.0
        department:
          type: string
          maxLength: 100
          minLength: 1
          title: Department
        enrollment_date:
          type: string
          format: date
          title: Enrollment Date
        status:
          $ref: '#/components/schemas/StudentStatus'
          default: active
        id:
          type: integer
          title: Id
        user_id:
          type: integer
          title: User Id
      type: object
      required:
      - university_id
      - name
      - department
      - id
      - user_id
      title: StudentResponse
    StudentStatus:
      type: string
      enum:
      - active
      - graduated
      - suspended
      title: StudentStatus
    StudentUpdate:
      properties:
        university_id:
          anyOf:
          - type: string
            maxLength: 20
            minLength: 3
          - type: 'null'
          title: University Id
        name:
          anyOf:
          - type: string
            maxLength: 150
            minLength: 1
          - type: 'null'
          title: Name
        birth_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Birth Date
        gender:
          anyOf:
          - $ref: '#/components/schemas/Gender'
          - type: 'null'
        phone_number:
          anyOf:
          - type: string
            maxLength: 20
          - type: 'null'
          title: Phone Number
        gpa:
          anyOf:
          - type: number
            maximum: 4.0
            minimum: 0.0
          - type: 'null'
          title: Gpa
        department:
          anyOf:
          - type: string
            maxLength: 100
            minLength: 1
          - type: 'null'
          title: Department
        enrollment_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Enrollment Date
        status:
          anyOf:
          - $ref: '#/components/schemas/StudentStatus'
          - type: 'null'
      type: object
      title: StudentUpdate
    Token:
      properties:
        access_token:
          type: string
          title: Access Token
        token_type:
          type: string
          title: Token Type
      type: object
      required:
      - access_token
      - token_type
      title: Token
    UserLogin:
      properties:
        email:
          type: string
          format: email
          title: Email
        password:
          type: string
          minLength: 1
          title: Password
      type: object
      required:
      - email
      - password
      title: UserLogin
    UserRegister:
      properties:
        email:
          type: string
          format: email
          title: Email
        password:
          type: string
          minLength: 1
          title: Password
        role:
          $ref: '#/components/schemas/Role'
      type: object
      required:
      - email
      - password
      - role
      title: UserRegister
    UserResponse:
      properties:
        id:
          type: integer
          title: Id
        email:
          type: string
          title: Email
        role:
          type: string
          title: Role
      type: object
      required:
      - id
      - email
      - role
      title: UserResponse
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /auth/login